feat(scripts): add fonts installation script

+++ scripts/fonts.sh
@@ -3,17 +3,26 @@
 echo "
 #######################
 # Fonts Installations #
-#######################
-"
+#######################"
 
-echo "Installing JetBrains Mono Nerd Font"
+echo ""
+echo "> Installing JetBrains Mono Nerd Font"
+echo ""
 
-mkdir -p ~/.local/share/fonts
+if ! grep -q "JetBrains Mono" ~/.local/share/fonts/*; then
+    echo "[!] JetBrains Mono Nerd Font is not installed. Installing now..."
+    echo ""
 
-wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/JetBrainsMono.zip
+    mkdir -p ~/.local/share/fonts
 
-unzip JetBrainsMono.zip -d ~/.local/share/fonts
+    wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/JetBrainsMono.zip
 
-rm JetBrainsMono.zip
+    unzip JetBrainsMono.zip -d ~/.local/share/fonts
 
-fc-cache -fv
+    rm JetBrainsMono.zip
+
+    fc-cache -fv
+else
+    echo "[!] JetBrains Mono Nerd Font is already installed."
+fi
+
